requirements.txt
torch
torchvision
numpy
pandas
matplotlib
seaborn
scikit-learn
evaluate
Pillow
argparse
